feat: create i18n sandbox page
